Moved `ExternalStakePool` file
