+++ src/views/Stake/Stake.tsx
@@ -37,12 +37,12 @@ import { error } from "../../slices/MessagesSlice";
 import { changeApproval, changeStake } from "../../slices/StakeThunk";
 import { changeApproval as changeGohmApproval } from "../../slices/WrapThunk";
 import { CurrentIndex, StakingAPY, TotalValueDeposited } from "../TreasuryDashboard/components/Metric/Metric";
+import ExternalStakePool from "./components/ExternalStakePool";
 import RebaseTimer from "./components/RebaseTimer/RebaseTimer";
 import { StakeFiveDayYield } from "./components/StakeFiveDayYield";
 import { StakeNextRebaseAmount } from "./components/StakeNextRebaseAmount";
 import { StakeRebaseYield } from "./components/StakeRebaseYield";
 import { ConfirmDialog } from "./ConfirmDialog";
-import ExternalStakePool from "./ExternalStakePool";
 
 const Stake: React.FC = () => {
   const dispatch = useDispatch();
