+++ src/views/Stake/components/ExternalStakePool.tsx
